+++ packages/components/histogram/src/HistogramG.svelte
@@ -320,6 +320,13 @@
 		selectedBins = [];
 		dispatch('clicked', {selectedBins});
 	}
+
+	const onKeyDown = event => {
+		if (event.key === 'Escape') {
+			event.preventDefault();
+			resetSelection();
+		}
+	}
 </script>
 
 <svelte:options namespace='svg' />
@@ -353,6 +360,7 @@
 					class:reset={selectedBins.length > 0}
 					class='bkgSensor'
 					on:click={resetSelection}
+					on:keydown={onKeyDown}
 				/>
 			{/if}