+++ packages/components/time_region_value/src/lib/components/GeoFilterModal.svelte
@@ -27,6 +27,7 @@
 <div
 	class='clickable modal'
 	on:click
+	on:keydown
 >
 	<div
 		class='panel'